refactor map view mv DevicesMap
