+++ demo-app/src/pages/platform-devices/MapView/DevicesMap/DevicesMap.jsx
