+++ demo-app/src/pages/platform-devices/MapView/DevicesMap/index.js
